protected routes with login required

+++ base/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -10,6 +12,9 @@ from .forms import RoomForm
 
 def login_page(request):
      
+     if request.user.is_authenticated:
+          return redirect('home')
+        
      if request.method == 'POST':
             username = request.POST.get('username')
             password = request.POST.get('password')
@@ -53,6 +58,7 @@ def room(request, id):
     context = {'room': room}
     return render(request, "base/room.html", context)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,11 +70,14 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def update_room(request, id):
         room = Room.objects.get(id=id)
         form = RoomForm(instance=room)
 
+        if request.user != room.host:
+             return HttpResponse('You are not allowed to do this action.')
+        
         if request.method == 'POST':
              form = RoomForm(request.POST, instance=room)
              if form.is_valid():
@@ -78,10 +87,14 @@ def update_room(request, id):
         context = {'form': form}
         return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def delete_room(request, id):
-     room = Room.objects.get(id=id)
-     if request.method == 'POST':
+    room = Room.objects.get(id=id)
+
+    if request.user != room.host:
+             return HttpResponse('You are not allowed to do this action.')
+     
+    if request.method == 'POST':
           room.delete()
           return redirect('home')
-     return render(request, 'base/delete.html', {'obj': room})
+    return render(request, 'base/delete.html', {'obj': room})
